Source_code/requirements.txt
plaintext
Pillow==8.4.0
tk==0.1.0
tkcolorpicker==2.2.2
